requirements.txt
pytest==8.0.0
python-dateutil==2.8.2
packaging==23.2
pluggy==1.4.0
iniconfig==2.0.0
exceptiongroup==1.2.0
pynacl==1.5.0
pycryptodome==3.18.0
requests==2.31.0
httpx==0.27.0
requests-toolbelt==0.7.1
pyrebase4==4.8.0
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
websockets==12.0
typing-extensions==4.9.0
dataclasses==0.6